pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "perpcity-sdk"
version = "0.4.2"
description = "Python SDK for interacting with Perp City perpetual futures contracts"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Strobe Labs" },
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries",
]
dependencies = [
    "web3>=7.0.0",
    "cachetools>=5.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "python-dotenv>=1.0",
    "mypy>=1.8",
    "ruff>=0.4",
]

[project.urls]
Homepage = "https://github.com/StrobeLabs/perpcity-python-sdk"
Repository = "https://github.com/StrobeLabs/perpcity-python-sdk"
Issues = "https://github.com/StrobeLabs/perpcity-python-sdk/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/perpcity_sdk"]

[tool.hatch.build.targets.sdist]
include = ["src/perpcity_sdk"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "integration: marks tests that require live blockchain access",
]
